Factoid fixes

+++ basement_bot/plugins/factoids.py
@@ -1,24 +1,29 @@
+import datetime
+
 from discord.ext import commands
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, DateTime, Integer, String
 
 from utils.cogs import MatchPlugin
-from utils.database import DatabaseHandler
+from utils.database import DatabaseHandle
 from utils.helpers import get_env_value, tagged_response
 
 FACTOID_PREFIX = get_env_value("FACTOID_PREFIX")
 COMMAND_PREFIX = get_env_value("COMMAND_PREFIX")
 
-db_handler = DatabaseHandler()
+db_handle = DatabaseHandle()
 
 
-class Factoid(db_handler.Base):
+class Factoid(db_handle.Base):
     __tablename__ = "factoids"
 
-    text = Column(String, primary_key=True)
+    pk = Column(Integer, primary_key=True)
+    text = Column(String)
+    channel = Column(String)
     message = Column(String)
+    time = Column(DateTime, default=datetime.datetime.utcnow)
 
 
-db_handler.create_all()
+db_handle.create_all()
 
 
 def setup(bot):
@@ -33,22 +38,32 @@ def setup(bot):
 
 @commands.command(name="r")
 async def add_factoid(ctx, arg1, *args):
+    if ctx.message.mentions:
+        await tagged_response(ctx, "Sorry, factoids don't work well with mentions.")
+        return
+
+    channel = str(ctx.message.channel.id)
+
     if not args:
         await tagged_response(ctx, "Factoids must not be blank!")
         return
 
-    db = db_handler.Session()
+    db = db_handle.Session()
 
     try:
         # first check if key already exists
-        entry = db.query(Factoid).filter(Factoid.text == arg1).first()
+        entry = (
+            db.query(Factoid)
+            .filter(Factoid.text == arg1, Factoid.channel == channel)
+            .first()
+        )
         if entry:
             # delete old one
             db.delete(entry)
             await tagged_response(ctx, "Deleting previous entry of factoid trigger...")
 
         # finally, add new entry
-        db.add(Factoid(text=arg1.lower(), message=" ".join(args)))
+        db.add(Factoid(text=arg1.lower(), channel=channel, message=" ".join(args)))
         db.commit()
         await tagged_response(ctx, f"Successfully added factoid trigger: *{arg1}*")
 
@@ -60,10 +75,20 @@ async def add_factoid(ctx, arg1, *args):
 
 @commands.command(name="f")
 async def delete_factoid(ctx, arg):
-    db = db_handler.Session()
+    if ctx.message.mentions:
+        await tagged_response(ctx, "Sorry, factoids don't work well with mentions.")
+        return
+
+    channel = str(ctx.message.channel.id)
+
+    db = db_handle.Session()
 
     try:
-        entry = db.query(Factoid).filter(Factoid.text == arg).first()
+        entry = (
+            db.query(Factoid)
+            .filter(Factoid.text == arg, Factoid.channel == channel)
+            .first()
+        )
         if entry:
             db.delete(entry)
             db.commit()
@@ -80,10 +105,20 @@ class FactoidMatch(MatchPlugin):
         return bool(content.startswith(FACTOID_PREFIX))
 
     async def response(self, ctx, arg):
-        db = db_handler.Session()
+        if ctx.message.mentions:
+            await tagged_response(ctx, "Sorry, factoids don't work well with mentions.")
+            return
+
+        channel = str(ctx.message.channel.id)
+
+        db = db_handle.Session()
 
         try:
-            entry = db.query(Factoid).filter(Factoid.text == arg[1:]).first()
+            entry = (
+                db.query(Factoid)
+                .filter(Factoid.text == arg[1:], Factoid.channel == channel)
+                .first()
+            )
             if entry:
                 await tagged_response(ctx, entry.message)
